Refactor emoji button
